+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## [Unreleased]
 ### Fixed
-- Corrected the Latvian translation of the phrase "Select terminal"
+- corrected the Latvian translation of the phrase "Select terminal"
+- fixed "Arrival email" additional service
+- added store email get from WP general info when no email is specified in plugin settings
+- disabled display of the return code in the delivery SMS
+- changed weight units to fixed value used in the plugin
+- added phone number service to courier-courier shipping
 
 ### Improved
 - added a ability to disable Omniva shipping methods for specific shipping class
+- improved additional services working
 
 ## [1.11.1]
 ### Fixed